Sketch out the source form
The presentation source will be a Markdown document embedded in a
minimal HTML skeleton. Slides will be separated with horizontal rules.
All other required code (scripts and styles) will eventually be in
separate files.

Note the choice of `<textarea>` element for the markdown source. This
is the only element that satisfies the following conditions:

1. is allowed in `<body>`, *and*
2. produces visible content with scripting disabled, *and*
3. allows (careful) use of unescaped `<` and `&`.

The latter property makes it possible to talk about HTML in
presentations without having to jump through too many hoops. The only
restriction is that we cannot mention `</textarea>`. When we need
that, we can escape `<` as `&lt;`.

Care is needed when `&` is followed by something that can be parsed as
an entity name. For example, `&region` will be parsed as `®ion` (even
though the semicolon is missing). In these cases, `&` has to be
escaped as `&amp;`.