Start with first slide by default
Since we only display the slide that is the target of the URL hash part,
there had better *be* a hash part. If there’s none, start with `#s1`.

Additionally, even if there is a hash part and I reload the page in
Firefox, it does not become the target initially. So force it.