Set font size
None of the usual font size units are good for presentations. A
comprehensive [research][1] on the matter involves taking into account
the screen size (when actually projected), the room depth (eye distance
of the farthest viewer), angular seconds and the visual acuity of
viewers.

[1]: http://www.thinkoutsidetheslide.com/selecting-the-correct-font-size/

We just base everything on an assumption of a maximum of 16 lines on
screen at 1.5 spacing.

100vh = 16 lines ⇒ 1 line = 6.25vh

6.25vh ÷ 1.5 = 4.17vh font size

# Please enter the commit message for your changes. Lines starting
# with '#' will be kept; you may remove them yourself if you want to.
# An empty message aborts the commit.
#
# Date:      Sun Jul 13 13:57:54 2014 +0700
#
# interactive rebase in progress; onto 189974e
# Last command done (1 command done):
#    reword dfed83b Set font size
# No commands remaining.
# You are currently editing a commit while rebasing branch 'tmp' on '189974e'.
#
# Changes to be committed:
#	modified:   lib/presenter.css
#	modified:   test/test.html
#
# Untracked files:
#	index.html~
#