Body text rhythm
A typical web page is scrolled continuously. In a presentation, on the
other side, one slide replaces another. It is important that the lines
appear at exact same positions.

We have a line height of 1.5. To completely preserve line rhythm, we
could:

* separate paragraphs with a margin of 1.5em; *or*
* indent the first line of each paragraph and do away with the margin.

Neither is satisfactory for presentations — 1.5em is too wide and
indentation plays poorly with lists.

So we choose to use a margin of 0.75em. This way, the rhythm breaks on
odd paragraphs, and is restored on even ones.