Add code highlighting
The README for `markdownit` has an example of using `highlight.js` but
it is synchronous. That is, every time `markdown-it` encounters a code
block, it stops and waits for highlighting to finish. In some cases,
that can take a few seconds, and we’re showing source to the user all
this time.

Therefore, we just let `markdown-it` do its job without any
highlighting. After it’s done, we load and initialize `highlight.js`.
With luck, it will have enough time to finish before the user gets to
the code blocks. Worst case, the user sees code unhighlighted.

Before actually kicking off the highlighting, we go through all code
blocks. Those that are tagged with a `language-*` class will be
highlighted according to that language, but for those that are not
tagged `highlight.js` will try to auto-detect the language. On blocks
in a typical presentation, auto-detection works unreliably. So we mark
them as `nohighlight`. Since `highlight.js` will skip such blocks
completely, we need to add the generic `hljs` class ourselves so these
blocks have common style with the others.

Finally, we adjust the styles. Monospaced fonts are wider than
proportional fonts are; if we keep normal size, code with moderately
long lines will not fit. But if we reduce the size to 80%, it is still
readable and fits 80+ columns on a 16:9 monitor.

To keep the line rhythm, we set line spacing to 150 / 80 = 1.875.
Also, the `highlight.js` style guide insists that all styles should
have a 0.5em padding on the `.hljs` class. This is misguided. We take
it out. Instead, we set 0.75em top and bottom margins on the `pre`.